+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     - Implement Carousel UI display with manual scroll [RMCCX-7619]
     - Implement user click behavior on Carousel Images [RMCCX-7618]
     - Implement autoscroll behavior for Carousel [RMCCX-7620]
+- Fixes:
+    - Fix issue of two images appearing on a single page in fullscreen Image Carousel campaign [RMCCX-8178]
 
 ### 9.0.0 (2024-10-22) 
 - Features: